requirements.txt
asgiref==3.9.1
boto3==1.40.14
botocore==1.40.14
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.7.2
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pandas==2.0.3
pillow==11.3.0
plotly==5.15.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.13.1
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
tenacity==9.1.2
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
